Code refactor
Refactored the codebase by merging overpayment.py and calculator.py.
