+++ mortgage/utils.py
@@ -30,49 +30,3 @@ def output_csv(func):
 
     return wrapper
 
-
-def loan_summary_decorator(summary_func):
-    """
-    Decorator to enhance the loan summary with additional details.
-    Includes overpayment details if applicable and comparison calculations if `compare` is True.
-    """
-
-    def wrapper(
-        self,
-        df: pd.DataFrame,
-        loan_amount: float,
-        overpayment_amount: float = 0,
-        compare: bool = False,
-    ):
-        # Generate the base summary
-        base_summary = summary_func(self, df, loan_amount, overpayment_amount)
-
-        # Dynamically retrieve overpayment_amount if applicable
-        overpayment_amount = getattr(self, "overpayment_amount", None)
-        if overpayment_amount is None and "Payment overpayment" in df.columns:
-            overpayment_amount = df["Payment overpayment"].iloc[0]
-
-        # Add comparison details if `compare` is True
-        if compare:
-            total_paid_ratio_overpayment = round(
-                df["Paid to date overpayment"].dropna().iloc[-1] / loan_amount, 2
-            )
-            total_years = df["Payment overpayment"].count() // 12
-            total_months = df["Payment overpayment"].count() % 12
-            interest_savings = round(
-                df["Interest charged to date standard"].iloc[-1]
-                - df["Interest charged to date overpayment"].dropna().iloc[-1],
-                2,
-            )
-            base_summary.append(
-                f"💸 Comparison Summary:"
-                f"\n   - Overpayment Repayment: £{total_paid_ratio_overpayment} per £1 borrowed"
-                f"\n   - Mortgage Term with Overpayment: {total_years} year(s) and {total_months} month(s)"
-                f"\n   - Interest Savings: £{interest_savings:,.2f}"
-            )
-
-        # Print the summary
-        print("\n".join(base_summary))
-        return base_summary
-
-    return wrapper
